move osc port mapping to own component. include that component on session page
